Fixed compatibility issue
Now after adding the app to home screen, user cannot zoom in the view. In addition, when click on the link, the app will not open the safari, instead, it will keep in the home screen page.

Some changes are made in Bootstrap.js. It might cause other side effect in the future